Backend/requirements.txt
# Основные зависимости
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
alembic==1.12.0
asyncpg==0.28.0
psycopg2-binary==2.9.7

# Безопасность
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Генерация отчетов
fpdf2==2.7.4
openpyxl==3.1.2

# Утилиты
python-dotenv==1.0.0
email-validator==2.0.0
httpx==0.24.1
